Move AuthenticationResponse to dto package
