+++ src/main/java/com/csemosip/bookingservice/service/Impl/AuthenticationServiceImpl.java
@@ -2,7 +2,7 @@ package com.csemosip.bookingservice.service.Impl;
 
 
 import com.csemosip.bookingservice.dto.AuthDTO;
-import com.csemosip.bookingservice.model.AuthenticationResponse;
+import com.csemosip.bookingservice.dto.AuthenticationResponse;
 
 public interface AuthenticationServiceImpl {
     AuthenticationResponse login(AuthDTO authDTO);
